requirements.txt
pandas
certifi
rfc5424-logging-handler
